Decouple GenerateTabs from SelectPlanPanel
